Completes login workflow

+++ ld33-common/src/main/java/com/seventh_root/ld33/common/network/packet/Packet.java
@@ -25,7 +25,7 @@ public abstract class Packet {
     public abstract int getId();
 
     public void write(ByteBuf buf) throws UnsupportedEncodingException {
-        buf.writeByte(getId());
+        buf.writeInt(getId());
     }
 
     public void writeString(ByteBuf buf, String string) throws UnsupportedEncodingException {
